JS removed debug window and use browser console instead
let debugbar indicate ajax errors; moved drag support to modal.js
