+++ wicket-devutils/src/main/java/org/apache/wicket/devutils/debugbar/DebugBar.html
@@ -21,10 +21,10 @@
 <body>
  
 	<wicket:panel>
-		<a id="wicketDebugBarCollapse" onclick="wicketDebugBarCollapse();"><img wicket:id="logo" src="wicket.png" alt="Wicket" /></a>
+		<a id="wicketDebugBarCollapse"><img wicket:id="logo" src="wicket.png" alt="Wicket" /></a>
 		<span id="wicketDebugBarContents" wicket:id="content">
 			<span wicket:id="contributors"><span wicket:id="contrib" class="contributor"></span></span>
-			<a id="wicketDebugBarRemove" onclick="wicketDebugBarRemove();"><img wicket:id="removeImg" src="remove.png" alt="Remove" /></a>
+			<a id="wicketDebugBarRemove"><img wicket:id="removeImg" src="remove.png" alt="Remove" /></a>
 		</span>
 	</wicket:panel>
